Use the data fetched by the loader in the Results Table
