+++ src/reducers/ComparisonSlice.ts
@@ -2,7 +2,6 @@ import { createSelector, createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 import { RootState } from '../common/store';
 import { Strings } from '../resources/Strings';
-import { CompareResultsItem, RevisionsHeader } from '../types/state';
 import { truncateHash } from '../utils/helpers';
 
 interface InitialState {
@@ -12,12 +11,6 @@ interface InitialState {
 const initialState: InitialState = {
   activeComparison:
     Strings.components.comparisonRevisionDropdown.allRevisions.key,
-};
-
-type Results = {
-  key: string;
-  value: CompareResultsItem[];
-  revisionHeader: RevisionsHeader;
 };
 
 type ResultsGroupedByKey = Record<string, ResultObject[]>;
@@ -105,63 +98,6 @@ export const selectStringifiedJsonResults = createSelector(
         null,
         2,
       );
-    }
-  },
-);
-
-function processResults(results: CompareResultsItem[]) {
-  const processedResults: Map<string, CompareResultsItem[]> = new Map<
-    string,
-    CompareResultsItem[]
-  >();
-  results.forEach((result) => {
-    const { new_rev: newRevision, header_name: header } = result;
-    const rowIdentifier = header.concat(' ', newRevision);
-    if (processedResults.has(rowIdentifier)) {
-      (processedResults.get(rowIdentifier) as CompareResultsItem[]).push(
-        result,
-      );
-    } else {
-      processedResults.set(rowIdentifier, [result]);
-    }
-  });
-  const restructuredResults: Results[] = Array.from(
-    processedResults,
-    function ([rowIdentifier, result]) {
-      return {
-        key: rowIdentifier,
-        value: result,
-        revisionHeader: {
-          suite: result[0].suite,
-          framework_id: result[0].framework_id,
-          test: result[0].test,
-          option_name: result[0].option_name,
-          extra_options: result[0].extra_options,
-          new_rev: result[0].new_rev,
-          new_repo: result[0].new_repository_name,
-        },
-      };
-    },
-  );
-
-  return restructuredResults;
-}
-
-const allRevisionsOption =
-  Strings.components.comparisonRevisionDropdown.allRevisions.key;
-
-export const selectProcessedResults = createSelector(
-  (state: RootState) => state.compareResults.data,
-  (state: RootState) => state.comparison.activeComparison,
-  (data, activeComparison) => {
-    if (activeComparison === allRevisionsOption) {
-      const allResults = ([] as CompareResultsItem[]).concat(
-        ...Object.values(data),
-      );
-      return processResults(allResults);
-    } else {
-      const results = data[activeComparison];
-      return processResults(results);
     }
   },
 );
